+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.1.0]
+
 ### Added
 - servers: allow adding server to a server group during server creation
 - servers: add server group UUID to server details
@@ -218,7 +220,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v5.0.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v5.1.0...HEAD
+[5.1.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v5.0.0...v5.1.0
 [5.0.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.10.0...v5.0.0
 [4.10.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.9.0...v4.10.0
 [4.9.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v4.8.0...v4.9.0